DEP combined all the classes
